% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityCause.R
\name{CauseBar}
\alias{CauseBar}
\title{Bar Chart showing the Causes of Driver Fatalities}
\usage{
CauseBar()
}
\value{
Bar Chart of Driver Fatality Causes
}
\description{
Creates a Bar Chart visualizing the causes of driver fatalities.
}
