% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityCause.R
\name{CauseGroupBar}
\alias{CauseGroupBar}
\title{Grouped Bar Chart showing the Causes of Driver Fatalities}
\usage{
CauseGroupBar()
}
\value{
Grouped Bar Chart of Driver Fatality Causes
}
\description{
Creates a Grouped Bar Chart visualizing the causes of driver fatalities grouped by year.
}
