% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityCause.R
\name{CauseLine}
\alias{CauseLine}
\title{Line Graph showing the Causes of Driver Fatalities}
\usage{
CauseLine()
}
\value{
Line Graph of Driver Fatality Causes
}
\description{
Creates a Line Graph visualizing the causes of driver fatalities.
}
