% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityCause.R
\name{CausePie}
\alias{CausePie}
\title{Pie Chart showing the Causes of Driver Fatalities}
\usage{
CausePie()
}
\value{
Pie Chart of Driver Fatality Causes
}
\description{
Creates a Pie Chart visualizing the causes of driver fatalities.
}
