% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityAge.R
\name{DriverAgeBar}
\alias{DriverAgeBar}
\title{Bar Chart showing Age of Driver Fatalities}
\usage{
DriverAgeBar()
}
\value{
Bar Chart of Driver Fatality Ages
}
\description{
Creates a Bar Chart visualizing the ages of driver fatalities.
}
