% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityAge.R
\name{DriverAgeGroupBar}
\alias{DriverAgeGroupBar}
\title{Grouped Bar Chart showing Age of Driver Fatalities}
\usage{
DriverAgeGroupBar()
}
\value{
Grouped Bar Chart of Driver Fatality Ages
}
\description{
Creates a Grouped Bar Chart visualizing the ages of driver fatalities grouped by year.
}
