% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityAge.R
\name{DriverAgeLine}
\alias{DriverAgeLine}
\title{Line Graph showing Age of Driver Fatalities}
\usage{
DriverAgeLine()
}
\value{
Line Graph of Driver Fatality Ages
}
\description{
Creates a Line Graph visualizing the ages of driver fatalities.
}
