% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityAge.R
\name{DriverAgePie}
\alias{DriverAgePie}
\title{Pie Chart showing Age of Driver Fatalities}
\usage{
DriverAgePie()
}
\value{
Pie Chart of Driver Fatality Ages
}
\description{
Creates a Pie Chart visualizing the ages of driver fatalities.
}
