% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityLoc.R
\name{FatalityPie}
\alias{FatalityPie}
\title{Pie Chart showing the Locations of Driver Fatalities}
\usage{
FatalityPie()
}
\value{
Pie Chart of Driver Fatality Locations
}
\description{
Creates a Pie Chart visualizing the Locations of driver fatalities.
}
