% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityLoc.R
\name{LocationLine}
\alias{LocationLine}
\title{Line Graph showing the Locations of Driver Fatalities}
\usage{
LocationLine()
}
\value{
Line Graph of Driver Fatality Locations
}
\description{
Creates a Line Graph visualizing the Locations of driver fatalities.
}
