% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityLoc.R
\name{LocationSAC}
\alias{LocationSAC}
\title{Stacked Area Chart showing the Locations of Driver Fatalities}
\usage{
LocationSAC()
}
\value{
Stack Area Chart of Driver Fatality Locations
}
\description{
Creates a Stacked Area Chart visualizing the Locations of driver fatalities.
}
