% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityVic.R
\name{VictimBar}
\alias{VictimBar}
\title{Bar Chart showing the Victims of Driver Fatalities}
\usage{
VictimBar()
}
\value{
Bar Chart of Driver Fatality Victims
}
\description{
Creates a Bar Chart visualizing the Victims of driver fatalities.
}
