% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityVic.R
\name{VictimGroupBar}
\alias{VictimGroupBar}
\title{Grouped Bar Chart showing the Victims of Driver Fatalities}
\usage{
VictimGroupBar()
}
\value{
Grouped Bar Chart of Driver Fatality Victims
}
\description{
Creates a Grouped Bar Chart visualizing the Victims of driver fatalities grouped by year.
}
