% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficFatalityVic.R
\name{VictimLine}
\alias{VictimLine}
\title{Line Graph showing the Victims of Driver Fatalities}
\usage{
VictimLine()
}
\value{
Line Graph of Driver Fatality Victims
}
\description{
Creates a Line Graph visualizing the Victims of driver fatalities.
}
